// packages/test/generatedCode/communications/calls/item/contentSharingSessions/item/contentSharingSessionItemRequestBuilderGetRequestConfiguration.ts
import {ContentSharingSessionItemRequestBuilderGetQueryParameters} from './contentSharingSessionItemRequestBuilderGetQueryParameters';
import {RequestOption} from '@microsoft/kiota-abstractions';

export interface ContentSharingSessionItemRequestBuilderGetRequestConfiguration {
    /** Request headers */
    headers?: Record<string, string>;
    /** Request options */
    options?: RequestOption[];
    /** Request query parameters */
    queryParameters?: ContentSharingSessionItemRequestBuilderGetQueryParameters;
}